moving nss implementation to its own file
